+++ changelog.md
@@ -1,11 +1,10 @@
 # Changelog
 
-* Item stats don't sync properly
-* Blast furnace ghost block don't work on servers
-
 ## Upcoming
 * Reduced mobs attack reach by ~15%
 * Respawn obelisks now have a higher chance to have iron blocks
+* Fixed Blast furnace ghost block not working on servers
+* Fixed item_stats data pack not loaded properly
 
 ## 1.1.2
 * Blast furnaces now show ghost blocks on how to build them